+++ js/util.js
@@ -1,0 +1,23 @@
+function getRandomInRange(min, max) {
+  if (typeof(min) === 'number' && typeof(max) === 'number') {
+    if (min >= 0 && max > 0) {
+      return (min < max) ? Math.floor(Math.random() * (max - min + 1)) + min : 'min больше или равен max';
+    }
+    return 'min или max не является положительным числом';
+  }
+  return 'min или max не является целым числом';
+}
+
+function checkMaxLengthString(string, length) {
+  return string.length < length;
+}
+
+const getRandomArrayElement = (elements) => {
+  const indexArray = elements[getRandomInRange(0, elements.length - 1)];
+  return indexArray;
+};
+
+getRandomInRange(0, 10);
+checkMaxLengthString('Привет', 10);
+
+export {getRandomInRange, checkMaxLengthString, getRandomArrayElement};
